+++ dimas/src/com/message.rs
@@ -14,29 +14,29 @@ use zenoh::{prelude::sync::SyncResolve, queryable::Query, sample::Sample};
 // region:		--- Message
 /// Implementation of a message received by subscriber callbacks
 #[derive(Debug)]
-pub struct Message {
-	/// the key expression on which the message was sent
-	pub key_expr: String,
-	/// the messages data
-	pub value: Vec<u8>,
-}
+pub struct Message(pub(crate)Sample);
 
 impl Deref for Message {
-	type Target = [u8];
+	type Target = Sample;
 
 	fn deref(&self) -> &Self::Target {
-		self.value.as_slice()
+		&self.0
 	}
 }
 
 impl Message {
 	/// decode message
 	/// # Errors
-	pub fn decode<T>(&self) -> Result<T, DimasError>
+	///
+	pub fn decode<T>(self) -> Result<T, DimasError>
 	where
 		T: Decode,
 	{
-		decode::<T>(self).map_err(|_| DimasError::DecodingFailed)
+		let value: Vec<u8> = self.0
+						.value
+						.try_into()
+						.map_err(|_| DimasError::DecodingFailed)?;
+		decode::<T>(value.as_slice()).map_err(|_| DimasError::DecodingFailed)
 	}
 }
 // endregion:	--- Message
@@ -44,24 +44,29 @@ impl Message {
 // region:    --- Request
 /// Implementation of a request for handling within a `Queryable`
 #[derive(Debug)]
-pub struct Request {
-	/// internal reference to zenoh `Query`
-	pub(crate) query: Query,
+pub struct Request(pub(crate) Query);
+
+impl Deref for Request {
+	type Target = Query;
+
+	fn deref(&self) -> &Self::Target {
+		&self.0
+	}
 }
 
 impl Request {
 	/// Reply to the given request
 	/// # Errors
 	///
-	pub fn reply<T>(&self, value: T) -> Result<(), DimasError>
+	pub fn reply<T>(self, value: T) -> Result<(), DimasError>
 	where
 		T: Encode,
 	{
-		let key = self.query.selector().key_expr.to_string();
+		let key = self.0.selector().key_expr.to_string();
 		let encoded: Vec<u8> = encode(&value).map_err(|_| DimasError::ShouldNotHappen)?;
 		let sample = Sample::try_from(key, encoded).map_err(|_| DimasError::ShouldNotHappen)?;
 
-		self.query
+		self.0
 			.reply(Ok(sample))
 			.res_sync()
 			.map_err(|_| DimasError::ShouldNotHappen)?;
@@ -71,7 +76,7 @@ impl Request {
 	/// access the queries parameters
 	#[must_use]
 	pub fn parameters(&self) -> &str {
-		self.query.parameters()
+		self.0.parameters()
 	}
 }
 // endregion: --- Request
@@ -79,29 +84,31 @@ impl Request {
 // region:		--- Response
 /// Implementation of a response received by query callbacks
 #[derive(Debug)]
-pub struct Response {
-	/// the key expression for which the response was sent
-	pub key_expr: String,
-	/// the responses data
-	pub value: Vec<u8>,
-}
+pub struct Response(pub(crate) Sample);
 
 impl Deref for Response {
-	type Target = [u8];
+	type Target = Sample;
 
 	fn deref(&self) -> &Self::Target {
-		self.value.as_slice()
+		&self.0
 	}
 }
 
 impl Response {
 	/// decode response
 	/// # Errors
-	pub fn decode<T>(&self) -> Result<T, DimasError>
+	///
+	/// # Panics
+	///
+	pub fn decode<T>(self) -> Result<T, DimasError>
 	where
 		T: Decode,
 	{
-		decode::<T>(self).map_err(|_| DimasError::DecodingFailed)
+		let value: Vec<u8> = self.0
+						.value
+						.try_into()
+						.map_err(|_| DimasError::DecodingFailed)?;
+		decode::<T>(value.as_slice()).map_err(|_| DimasError::DecodingFailed)
 	}
 }
 // endregion:	--- Response