+++ dimas/src/com/query.rs
@@ -18,7 +18,7 @@ use zenoh::{
 #[allow(clippy::module_name_repetitions)]
 pub type QueryCallback<P> = Arc<
 	Mutex<
-		dyn FnMut(&ArcContext<P>, &Response) -> Result<(), DimasError>
+		dyn FnMut(&ArcContext<P>, Response) -> Result<(), DimasError>
 			+ Send
 			+ Sync
 			+ Unpin
@@ -72,7 +72,7 @@ where
 	#[must_use]
 	pub fn callback<F>(mut self, callback: F) -> Self
 	where
-		F: FnMut(&ArcContext<P>, &Response) -> Result<(), DimasError>
+		F: FnMut(&ArcContext<P>, Response) -> Result<(), DimasError>
 			+ Send
 			+ Sync
 			+ Unpin
@@ -174,25 +174,15 @@ where
 			//.timeout(Duration::from_millis(1000))
 			.res_sync()
 			.expect("should never happen");
-		//dbg!(&replies);
 
 		while let Ok(reply) = replies.recv() {
-			//dbg!(&reply);
 			match reply.sample {
 				Ok(sample) => {
-					//dbg!(&sample);
-					let value: Vec<u8> = sample
-						.value
-						.try_into()
-						.expect("should not happen");
-					let msg = Response {
-						key_expr: sample.key_expr.to_string(),
-						value,
-					};
 					match sample.kind {
 						SampleKind::Put => {
+							let msg = Response(sample);
 							if let Err(error) =
-								cb.lock().expect("should not happen")(&self.ctx, &msg)
+								cb.lock().expect("should not happen")(&self.ctx, msg)
 							{
 								error!("call failed with {error}");
 							};