+++ examples/queries/src/query.rs
@@ -21,7 +21,7 @@ struct Args {
 #[derive(Debug)]
 struct AgentProps {}
 
-fn query_callback(_ctx: &ArcContext<AgentProps>, response: &Response) -> Result<(), DimasError> {
+fn query_callback(_ctx: &ArcContext<AgentProps>, response: Response) -> Result<(), DimasError> {
 	let message: u128 = response.decode()?;
 	println!("Response is '{message}'");
 	Ok(())